refactor: 참여자명단 초기화 작업 Service에게
String 형태의 참가자 명단을 Car 객체로 담아 저장하는 로직을 Service에서 수행하도록 변경
