+++ src/main/java/racingcar/car/model/CarRepositoryImpl.java
@@ -3,7 +3,6 @@ package racingcar.car.model;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.StringTokenizer;
 
 import racingcar.car.entity.Car;
 
@@ -21,8 +20,8 @@ public class CarRepositoryImpl implements CarRepository {
 	}
 
 	@Override
-	public void saveCar(String initCars) {
-		splitInitCar(initCars);
+	public void saveCar(Car newCar) {
+		cars.add(newCar);
 	}
 
 	@Override
@@ -33,13 +32,5 @@ public class CarRepositoryImpl implements CarRepository {
 	@Override
 	public void setPosition() {
 		cars.forEach(Car::randomPosition);
-	}
-
-	private void splitInitCar(String initCars) {
-		StringTokenizer stringTokenizer = new StringTokenizer(initCars, ",");
-		while (stringTokenizer.hasMoreTokens()) {
-			Car newCar = new Car(stringTokenizer.nextToken());
-			cars.add(newCar);
-		}
 	}
 }
